pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compact-memory"
version = "1.0.0"
description = "Prototype implementation of the Compact Memory Agent"
authors = [{name = "Scott Falconer", email = "scottfalconer@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
dependencies = [
    "openai",
    "tiktoken",
    "numpy",
    "faiss-cpu",
    "click>=8.2",
    "tqdm",
    "pydantic",
    "pyyaml",
    "typer[all]>=0.16.0",
    "portalocker",
    "rich>=13.6",
]

[project.optional-dependencies]
test = ["pytest"]
chroma = ["chromadb"]
spacy = ["spacy"]
optuna = ["optuna"]
ray = ["ray[tune]"]
embedding = ["sentence-transformers", "torch"]
local = ["transformers", "torch"]
gemini = ["google-generativeai"]
metrics = ["evaluate"]

[project.scripts]
compact-memory = "compact_memory.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["compact_memory", "compact_memory.*", "CompressionEngine", "CompressionEngine.*"]

[tool.setuptools.package-data]
"compact_memory" = ["py.typed"]
